Added cli file and moved questions there
