+++ README.md
@@ -1,1 +1,3 @@
 # Employee-tracker-mysql
+
+Adde the environment variables .env.example in the repo
