+++ routes/employees.js
@@ -6,13 +6,13 @@ const db = require('../config/conecction');
 
 router.get('/', (req, res) => {
   const sql = `SELECT id, first_name, last_name FROM employees`;
-  db.query(sql, (err, emmployees) => {
+  db.query(sql, (err, employees) => {
     if (err) {
       return res.status(500).json({ message: 'error', error: err.message });
     }
     res.json({
       message: 'success',
-      data: emmployees
+      data: employees
     });
   });  
 });
